+++ src/client/test/Integration/power-pages/actions-hub/tree-items/ActionsHubTreeItem.test.ts
@@ -5,7 +5,7 @@
 
 import * as vscode from "vscode";
 import { expect } from "chai";
-import { ActionsHubTreeItem } from "../../../../../../power-pages/actions-hub/tree-items/ActionsHubTreeItem";
+import { ActionsHubTreeItem } from "../../../../../power-pages/actions-hub/tree-items/ActionsHubTreeItem";
 
 class MockTreeItem extends ActionsHubTreeItem {
     constructor() {
